Remove some warning.

+++ CPPSources/Utils/FileContent.h
@@ -3,7 +3,7 @@
 
 struct FileContent{
     
-    FileContent(char* contentChars, unsigned int cSize){ content = contentChars; size = cSize;}
+    FileContent(char* contentChars, long cSize){ content = contentChars; size = cSize;}
     ~FileContent(){ if(content) delete[] content;
 //        logMessage("FILE CONTENT DESTRUCTOR!");
     }
